moved new upsample test to upsample nightly
